# tequilapi/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this documentation is to provide developers an insight of how to\ninteract with Mysterium Node via Tequila API.\nThis should demonstrate all the possible API calls with described parameters and responses.",
    "title": "Tequila API",
    "version": "dev"
  },
  "host": "127.0.0.1:4050",
  "paths": {
...
    "/proposals": {
      "get": {
        "description": "Returns list of proposals filtered by provider id",
        "tags": [
          "Proposal"
        ],
        "summary": "Returns proposals",
        "operationId": "listProposals",
        "parameters": [
          {
            "type": "string",
            "description": "id of provider proposals",
            "name": "provider_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the service type of the proposal. Possible values are \"openvpn\", \"wireguard\" and \"noop\"",
            "name": "service_type",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the access policy id to filter the proposals by",
            "name": "access_policy",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the access policy source to filter the proposals by",
            "name": "access_policy_source",
            "in": "query"
          },
          {
            "type": "string",
            "description": "If given will filter proposals by node location country.",
            "name": "country",
            "in": "query"
          },
          {
            "type": "string",
            "description": "IP Type (residential, datacenter, etc.).",
            "name": "ip_type",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Minimum compatibility level of the proposal.",
            "name": "compatibility_min",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Maximum compatibility level of the proposal.",
            "name": "compatibility_max",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Minimum quality of the provider.",
            "name": "quality_min",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Pick nodes compatible with NAT of specified type. Specify \"auto\" to probe NAT.",
            "name": "nat_compatibility",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of proposals",
            "schema": {
              "$ref": "#/definitions/ListProposalsResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/proposals/countries": {
      "get": {
        "description": "Returns a list of countries with a number of proposals",
        "tags": [
          "Countries"
        ],
        "summary": "Returns number of proposals per country",
        "operationId": "listCountries",
        "parameters": [
          {
            "type": "string",
            "description": "id of provider proposals",
            "name": "provider_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the service type of the proposal. Possible values are \"openvpn\", \"wireguard\" and \"noop\"",
            "name": "service_type",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the access policy id to filter the proposals by",
            "name": "access_policy",
            "in": "query"
          },
          {
            "type": "string",
            "description": "the access policy source to filter the proposals by",
            "name": "access_policy_source",
            "in": "query"
          },
          {
            "type": "string",
            "description": "If given will filter proposals by node location country.",
            "name": "country",
            "in": "query"
          },
          {
            "type": "string",
            "description": "IP Type (residential, datacenter, etc.).",
            "name": "ip_type",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Minimum compatibility level of the proposal.",
            "name": "compatibility_min",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Maximum compatibility level of the proposal.",
            "name": "compatibility_max",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Minimum quality of the provider.",
            "name": "quality_min",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Pick nodes compatible with NAT of specified type. Specify \"auto\" to probe NAT.",
            "name": "nat_compatibility",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of countries",
            "schema": {
              "$ref": "#/definitions/ListProposalsCountiesResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/proposals/filter-presets": {
      "get": {
        "description": "Returns proposal filter presets",
        "tags": [
          "Proposal"
        ],
        "summary": "Returns proposal filter presets",
        "operationId": "proposalFilterPresets",
        "responses": {
          "200": {
            "description": "List of proposal filter presets",
            "schema": {
              "$ref": "#/definitions/ListProposalFilterPresetsResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "APIError": {
      "type": "object",
      "title": "APIError represents an error response from REST API service.",
      "properties": {
        "error": {
          "$ref": "#/definitions/Err"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "apiErrorSwagger",
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "AccessPolicy": {
      "description": "AccessPolicy represents the access controls for proposal",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/market"
    },
...
    "Err": {
      "type": "object",
      "title": "Err contains the error details.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "fields": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/FieldError"
          },
          "x-go-name": "Fields"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "FieldError": {
      "type": "object",
      "title": "FieldError contains the reason why a field failed validation.",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
    "FilterPreset": {
      "type": "object",
      "title": "FilterPreset is a pre-defined proposal filter.",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ListProposalFilterPresetsResponse": {
      "type": "object",
      "title": "ListProposalFilterPresetsResponse holds a list of proposal filter presets.",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FilterPreset"
          },
          "x-go-name": "Items"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ListProposalsCountiesResponse": {
      "type": "object",
      "title": "ListProposalsCountiesResponse holds number of proposals per country.",
      "additionalProperties": {
        "type": "integer",
        "format": "int64"
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ListProposalsResponse": {
      "type": "object",
      "title": "ListProposalsResponse holds list of proposals.",
      "properties": {
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProposalDTO"
          },
          "x-go-name": "Proposals"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "Price": {
      "type": "object",
      "title": "Price represents the service price.",
      "properties": {
        "currency": {
          "type": "string",
          "x-go-name": "Currency"
        },
        "per_gib": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "PerGiB"
        },
        "per_gib_tokens": {
          "$ref": "#/definitions/Tokens"
        },
        "per_hour": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "PerHour"
        },
        "per_hour_tokens": {
          "$ref": "#/definitions/Tokens"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ProposalDTO": {
      "type": "object",
      "title": "ProposalDTO holds service proposal details.",
      "properties": {
        "access_policies": {
          "description": "AccessPolicies",
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccessPolicy"
          },
          "x-go-name": "AccessPolicies"
        },
        "compatibility": {
          "description": "Compatibility level.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Compatibility"
        },
        "format": {
          "description": "Proposal format.",
          "type": "string",
          "x-go-name": "Format"
        },
        "location": {
          "$ref": "#/definitions/ServiceLocationDTO"
        },
        "price": {
          "$ref": "#/definitions/Price"
        },
        "provider_id": {
          "description": "provider who offers service",
          "type": "string",
          "x-go-name": "ProviderID",
          "example": "0x0000000000000000000000000000000000000001"
        },
        "quality": {
          "$ref": "#/definitions/Quality"
        },
        "service_type": {
          "description": "type of service provider offers",
          "type": "string",
          "x-go-name": "ServiceType",
          "example": "openvpn"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "Quality": {
      "type": "object",
      "title": "Quality holds proposal quality metrics.",
      "properties": {
        "bandwidth": {
          "type": "number",
          "format": "double",
          "x-go-name": "Bandwidth"
        },
        "latency": {
          "type": "number",
          "format": "double",
          "x-go-name": "Latency"
        },
        "quality": {
          "type": "number",
          "format": "double",
          "x-go-name": "Quality"
        },
        "uptime": {
          "type": "number",
          "format": "double",
          "x-go-name": "Uptime"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ServiceLocationDTO": {
      "type": "object",
      "title": "ServiceLocationDTO holds service location metadata.",
      "properties": {
        "asn": {
          "description": "Autonomous System Number",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ASN",
          "example": 1
        },
        "city": {
          "type": "string",
          "x-go-name": "City",
          "example": "Amsterdam"
        },
        "continent": {
          "type": "string",
          "x-go-name": "Continent",
          "example": "EU"
        },
        "country": {
          "type": "string",
          "x-go-name": "Country",
          "example": "NL"
        },
        "ip_type": {
          "type": "string",
          "x-go-name": "IPType",
          "example": "residential"
        },
        "isp": {
          "type": "string",
          "x-go-name": "ISP",
          "example": "Telia Lietuva, AB"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "Tokens": {
      "description": "Tokens a common response for ethereum blockchain monetary amount",
      "type": "object",
      "properties": {
        "ether": {
          "type": "string",
          "x-go-name": "Ether"
        },
        "human": {
          "type": "string",
          "x-go-name": "Human"
        },
        "wei": {
          "type": "string",
          "x-go-name": "Wei"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
  }
}